Add contiguous_strides as a correct replacement of defaultStride
`at::defaultStride` was added in
https://github.com/pytorch/pytorch/pull/18779
As it was noted in that PR, it differs from the actual computation of
the default strides when one or more of the dimensions of the tensor are
zero. See
https://github.com/pytorch/pytorch/pull/18779#discussion_r272296140

We add two functions, `contiguous_strides` and `contiguous_strides_vec`
which correct this issue and we replace the previous (wrong) uses of
`defaultStride`.

ghstack-source-id: 0023f20cac5d9638fe00654c9158ef41951705bb
Pull Request resolved: https://github.com/pytorch/pytorch/pull/67789